Created RestaurantListContainer to handle mapping over data
