+++ src/components/RestaurantsNear/header.test.js
@@ -1,5 +1,6 @@
 import React from "react";
 import renderer from "react-test-renderer";
+
 import { RestaurantsNear } from "./index";
 
 it("displays the Restaurants Near Header correctly", () => {
