+++ src/containers/RestaurantListContainer/restaurantListContainerUtils.js
@@ -1,0 +1,13 @@
+import { LabelsObject } from "../../Labels";
+
+const checkValidImageURL = (url) => {
+  const { invalidImageURLLabel } = LabelsObject;
+
+  if (url === "https://s3-media3.fl.yelpcdn.com/bphoto/None/o.jpg") {
+    return invalidImageURLLabel;
+  } else {
+    return url;
+  }
+};
+
+export { checkValidImageURL };
