+++ src/containers/SearchResults/searchResultsUtils.js
@@ -36,14 +36,4 @@ const fetchRestaurantData = async (
   updateLoadingState(false);
 };
 
-const checkValidImageURL = (url) => {
-  const { invalidImageURLLabel } = LabelsObject;
-
-  if (url === "https://s3-media3.fl.yelpcdn.com/bphoto/None/o.jpg") {
-    return invalidImageURLLabel;
-  } else {
-    return url;
-  }
-};
-
-export { fetchRestaurantData, checkValidImageURL };
+export { fetchRestaurantData };
